elf: Move ElfError to separate module
In the effort of splitting up the ELF parser code, move the ElfError
implementation to its own module.
